Move Session Config up
