+++ core/src/main/java/io/opentelemetry/android/internal/session/SessionIdTimeoutHandler.kt
@@ -6,7 +6,7 @@
 package io.opentelemetry.android.internal.session
 
 import io.opentelemetry.android.internal.services.applifecycle.ApplicationStateListener
-import io.opentelemetry.android.session.SessionConfig
+import io.opentelemetry.android.config.SessionConfig
 import io.opentelemetry.sdk.common.Clock
 import kotlin.time.Duration
